+++ tests/integration_tests/adapters/binance/resources/http_wallet_sandbox.py
@@ -19,27 +19,28 @@ import os
 
 import pytest
 
-from nautilus_trader.adapters.binance.http.api.spot import BinanceSpotHTTPAPI
-from nautilus_trader.adapters.binance.http.client import BinanceHttpClient
+from nautilus_trader.adapters.binance.factories import get_cached_binance_http_client
+from nautilus_trader.adapters.binance.http.api.wallet import BinanceWalletHttpAPI
 from nautilus_trader.common.clock import LiveClock
 from nautilus_trader.common.logging import Logger
 
 
 @pytest.mark.asyncio
-async def test_binance_http_client():
+async def test_binance_spot_wallet_http_client():
     loop = asyncio.get_event_loop()
     clock = LiveClock()
 
-    client = BinanceHttpClient(
+    client = get_cached_binance_http_client(
         loop=loop,
         clock=clock,
         logger=Logger(clock=clock),
         key=os.getenv("BINANCE_API_KEY"),
         secret=os.getenv("BINANCE_API_SECRET"),
-        base_url="https://api.binance.com",
     )
 
-    market = BinanceSpotHTTPAPI(client=client)
+    wallet = BinanceWalletHttpAPI(client=client)
     await client.connect()
-    response = await market.depth("ETHUSDT")
+    response = await wallet.trade_fee(symbol="BTCUSDT")
     print(json.dumps(json.loads(response), indent=4))
+
+    await client.disconnect()