+++ tests/test_kit/mocks.py
@@ -25,6 +25,7 @@ from fsspec.implementations.memory import MemoryFileSystem
 from nautilus_trader.accounting.accounts.base import Account
 from nautilus_trader.cache.database import CacheDatabase
 from nautilus_trader.common.actor import Actor
+from nautilus_trader.common.config import ActorConfig
 from nautilus_trader.common.logging import Logger
 from nautilus_trader.common.providers import InstrumentProvider
 from nautilus_trader.core.datetime import secs_to_nanos
@@ -116,12 +117,12 @@ class MockActor(Actor):
     Provides a mock actor for testing.
     """
 
-    def __init__(self):
+    def __init__(self, config: ActorConfig = None):
         """
         Initialize a new instance of the ``MockActor`` class.
 
         """
-        super().__init__()
+        super().__init__(config)
 
         self.object_storer = ObjectStorer()
 
@@ -570,8 +571,8 @@ class MockLiveExecutionClient(LiveExecutionClient):
             logger=logger,
         )
 
-        self._order_status_reports = {}  # type: dict[VenueOrderId, OrderStatusReport]
-        self._trades_lists = {}  # type: dict[VenueOrderId, list[ExecutionReport]]
+        self._order_status_reports: Dict[VenueOrderId, OrderStatusReport] = {}
+        self._trades_lists: Dict[VenueOrderId, list[ExecutionReport]] = {}
 
         self.calls = []
         self.commands = []
@@ -641,11 +642,11 @@ class MockCacheDatabase(CacheDatabase):
         """
         super().__init__(logger)
 
-        self.currencies = {}  # type: dict[str, Currency]
-        self.instruments = {}  # type: dict[InstrumentId, Instrument]
-        self.accounts = {}  # type: dict[AccountId, Account]
-        self.orders = {}  # type: dict[ClientOrderId, Order]
-        self.positions = {}  # type: dict[PositionId, Position]
+        self.currencies: Dict[str, Currency] = {}
+        self.instruments: Dict[InstrumentId, Instrument] = {}
+        self.accounts: Dict[AccountId, Account] = {}
+        self.orders: Dict[ClientOrderId, Order] = {}
+        self.positions: Dict[PositionId, Position] = {}
 
     def flush(self) -> None:
         self.accounts.clear()